+++ src/Mac.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Security;
 
+use SensitiveParameter;
 use Yiisoft\Strings\StringHelper;
 
 /**
@@ -51,7 +52,12 @@ final class Mac
      * @see hkdf()
      * @see pbkdf2()
      */
-    public function sign(string $data, #[\SensitiveParameter] string $key, bool $rawHash = false): string
+    public function sign(
+        string $data,
+        #[SensitiveParameter]
+        string $key,
+        bool $rawHash = false
+    ): string
     {
         $hash = hash_hmac($this->algorithm, $data, $key, $rawHash);
         if (!$hash) {
@@ -80,7 +86,12 @@ final class Mac
      *
      * @see hash()
      */
-    public function getMessage(string $data, #[\SensitiveParameter] string $key, bool $rawHash = false): string
+    public function getMessage(
+        string $data,
+        #[SensitiveParameter]
+        string $key,
+        bool $rawHash = false
+    ): string
     {
         $test = hash_hmac($this->algorithm, '', '', $rawHash);
         if (!$test) {