+++ src/PasswordHasher.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\Security;
 
+use SensitiveParameter;
+
 /**
  * PasswordHasher allows generating password hash and verifying passwords against a hash.
  */
@@ -65,7 +67,10 @@ final class PasswordHasher
      * @psalm-suppress InvalidNullableReturnType
      * @psalm-suppress NullableReturnStatement
      */
-    public function hash(#[\SensitiveParameter] string $password): string
+    public function hash(
+        #[SensitiveParameter]
+        string $password
+    ): string
     {
         return password_hash($password, $this->algorithm, $this->parameters);
     }
@@ -83,7 +88,11 @@ final class PasswordHasher
      *
      * @see hash()
      */
-    public function validate(#[\SensitiveParameter] string $password, string $hash): bool
+    public function validate(
+        #[SensitiveParameter]
+        string $password,
+        string $hash
+    ): bool
     {
         if ($password === '') {
             throw new \InvalidArgumentException('Password must be a string and cannot be empty.');